+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Notable changes to this project.
 ## dev
 - signals: support upload to diagnostic tool
 - cli: added `list-datasets`
+- cli: implement downloading of the new dataset
 - some code cleanup
 
 ## [2.9.1] - 2021-09-27